DryveBackend/requirements.txt
requests>=2.21.0
Flask>=1.0.2
google-auth>=1.5.1
googlemaps>=3.0.2
google-cloud-firestore>=0.31.0